Moved collection methods into their own 'CollectionController' in order to isolate for testing. Refactred views and corrsponding links appropriately to function with this new controller.
