+++ main_project_code/TeamProject/iCollections/Data/Abstract/ICollectionRepository.cs
@@ -5,10 +5,5 @@ namespace iCollections.Data.Abstract
     public interface ICollectionRepository : IRepository<Collection>
     {
         Collection GetIcollectionsByUsername(string username);
-        bool Exists(IcollectionUser user);
-
-        int GetReadableUserID(string username);
-
-        int GetProfilePicID(int userId);
     }
 }
