+++ main_project_code/TeamProject/iCollections/Startup.cs
@@ -48,6 +48,7 @@ namespace iCollections
             services.AddDatabaseDeveloperPageExceptionFilter();
 
             services.AddScoped<IPhotoRepository, PhotoRepository>();
+            services.AddScoped<ICollectionRepository, CollectionRepository>();
             services.AddScoped<IIcollectionUserRepository, IcollectionUserRepository>();
 
             services.AddDefaultIdentity<IdentityUser>(options => options.SignIn.RequireConfirmedAccount = true)
